backend_service/requirements.txt
flask
psycopg2-binary
influxdb
schedule
